requirements.txt
PyMuPDF
pytesseract
pdf2image
python-dotenv
regex
pandas
scikit-learn
joblib
argparse
transformers
